+++ projects/jrng-ui/speech-to-text/public-api.ts
@@ -1,0 +1,3 @@
+export * from './speech-recognition.service';
+export * from './speech-to-text.directive';
+export * from './speech-to-text-button.component';